mv CMath to a new file
